refactor(@deip/attributes-module): add attributes helpers and remove platform-fns
