+++ .eslintignore
@@ -6,5 +6,5 @@ packages/common/lib-crypto/**
 packages/casimir/services/RpcClient/**
 packages/casimir/services/Chain/**
 packages/casimir/services/Grants/**
-packages/common/platform-fns/lib/helpers/_temp.js
+packages/casimir/modules/attributes/lib/composables/_temp.js
 packages/generators/**/generators/*/templates/**
