+++ packages/casimir/modules/assets/package.json
@@ -19,7 +19,6 @@
 	"dependencies": {
 		"@deip/assets-service": "1.366.0",
 		"@deip/constants": "1.366.0",
-		"@deip/platform-fns": "1.366.0",
 		"@deip/platform-store": "1.366.0",
 		"@deip/platform-util": "1.366.0",
 		"@deip/proxydi": "1.366.0",
