+++ packages/casimir/modules/attributes/index.js
@@ -7,3 +7,4 @@ export { default as AttributeRead } from './lib/components/AttributeRead';
 export { default as AttributesSetup } from './lib/components/AttributesSetup';
 
 export { AttributesList } from './lib/components/AttributesList';
+export * from './lib/composables/attributes';
