+++ packages/casimir/modules/auth/package.json
@@ -24,7 +24,6 @@
     "@deip/auth-service": "1.366.0",
     "@deip/constants": "1.366.0",
     "@deip/lib-crypto": "1.365.0",
-    "@deip/platform-fns": "1.366.0",
     "@deip/platform-store": "1.366.0",
     "@deip/proxydi": "1.366.0",
     "@deip/toolbox": "1.366.0",
