+++ packages/casimir/modules/domains/package.json
@@ -25,7 +25,6 @@
   "dependencies": {
     "@deip/constants": "1.366.0",
     "@deip/domains-service": "1.366.0",
-    "@deip/platform-fns": "1.366.0",
     "@deip/platform-store": "1.366.0",
     "@deip/proxydi": "1.366.0",
     "@deip/toolbox": "1.366.0",
