+++ packages/casimir/modules/fundraising/package.json
@@ -27,7 +27,6 @@
     "@deip/constants": "1.366.0",
     "@deip/investments-service": "1.366.0",
     "@deip/platform-components": "1.366.0",
-    "@deip/platform-fns": "1.366.0",
     "@deip/platform-store": "1.366.0",
     "@deip/platform-util": "1.366.0",
     "@deip/projects-module": "1.366.0",
