+++ packages/casimir/modules/layouts/package.json
@@ -22,7 +22,6 @@
     "@deip/constants": "1.366.0",
     "@deip/layout-service": "1.366.0",
     "@deip/platform-components": "1.366.0",
-    "@deip/platform-fns": "1.366.0",
     "@deip/platform-store": "1.366.0",
     "@deip/platform-util": "1.366.0",
     "@deip/proxydi": "1.366.0",
