+++ packages/casimir/modules/projects/package.json
@@ -24,7 +24,6 @@
   },
   "dependencies": {
     "@deip/constants": "1.366.0",
-    "@deip/platform-fns": "1.366.0",
     "@deip/platform-store": "1.366.0",
     "@deip/project-service": "1.366.0",
     "@deip/proxydi": "1.366.0",
