+++ packages/casimir/modules/teams/package.json
@@ -24,7 +24,6 @@
     "@deip/constants": "1.366.0",
     "@deip/layouts-module": "1.366.0",
     "@deip/platform-components": "1.366.0",
-    "@deip/platform-fns": "1.366.0",
     "@deip/platform-store": "1.366.0",
     "@deip/proxydi": "1.366.0",
     "@deip/team-service": "1.366.0",
