+++ packages/casimir/modules/tenants/package.json
@@ -21,7 +21,6 @@
   },
   "dependencies": {
     "@deip/constants": "1.366.0",
-    "@deip/platform-fns": "1.366.0",
     "@deip/platform-store": "1.366.0",
     "@deip/portal-service": "1.366.0",
     "@deip/proxydi": "1.366.0"
