+++ packages/casimir/modules/users/package.json
@@ -25,7 +25,6 @@
     "@deip/constants": "1.366.0",
     "@deip/layouts-module": "1.366.0",
     "@deip/platform-components": "1.366.0",
-    "@deip/platform-fns": "1.366.0",
     "@deip/platform-store": "1.366.0",
     "@deip/platform-util": "1.366.0",
     "@deip/proxydi": "1.366.0",
