+++ packages/casimir/platform/components/package.json
@@ -21,7 +21,6 @@
   },
   "dependencies": {
     "@deip/constants": "1.366.0",
-    "@deip/platform-fns": "1.366.0",
     "@deip/platform-util": "1.366.0",
     "@deip/toolbox": "1.366.0",
     "@deip/vuetify-extended": "1.366.0",
