--- packages/common/platform-fns/README.md
@@ -1,11 +1,0 @@
-# `@deip/platform-fns`
-
-> TODO: description
-
-## Usage
-
-```
-const platformFns = require('@deip/platform-fns');
-
-// TODO: DEMONSTRATE API
-```
