--- packages/common/platform-fns/index.js
@@ -1,1 +1,0 @@
-export * from './lib/helpers';
