--- packages/common/platform-fns/lib/helpers/index.js
@@ -1,1 +1,0 @@
-export * from './attributes';
