--- packages/common/platform-fns/package.json
@@ -1,42 +1,0 @@
-{
-  "name": "@deip/platform-fns",
-  "version": "1.366.0",
-  "description": "> TODO: description",
-  "author": "Euheny Bondarovich <euheny@gmail.com>",
-  "homepage": "https://gitlab.com/DEIP/deip-client-modules#readme",
-  "license": "ISC",
-  "main": "index.js",
-  "publishConfig": {
-    "registry": "https://npm-registry.deip.world/"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+ssh://git@gitlab.com/DEIP/deip-client-modules.git"
-  },
-  "scripts": {
-    "test": "echo \"Error: run tests from root\" && exit 1"
-  },
-  "bugs": {
-    "url": "https://gitlab.com/DEIP/deip-client-modules/issues"
-  },
-  "dependencies": {
-    "@deip/constants": "1.366.0",
-    "@deip/proxydi": "1.366.0",
-    "@deip/toolbox": "1.366.0",
-    "change-case": "^4.1.2",
-    "date-fns": "^2.22.1",
-    "qs": "^6.10.1"
-  },
-  "peerDependencies": {
-    "vue": "^2.6.14",
-    "vuetify": "^2.5.3"
-  },
-  "devDependencies": {
-    "@babel/core": "^7.14.3",
-    "@babel/plugin-proposal-class-properties": "^7.13.0",
-    "@babel/plugin-proposal-decorators": "^7.14.2",
-    "@babel/plugin-proposal-object-rest-spread": "^7.14.4",
-    "@babel/plugin-proposal-optional-chaining": "^7.14.2",
-    "@babel/plugin-proposal-private-methods": "^7.13.0"
-  }
-}
